insert new Subset records into the database
includes logic to retrieve existing Subset records, insert new ones, associate Subset to both CardSet and Card in-memory, new queries to Common Queries SQL file, and an update in parsing logic to set the grading modifier field to 1.0 by default. Card records in the database are NOT associated to new Subset records yet.